+++ src/main/cfg_mgr.c
@@ -695,10 +695,10 @@ cfg_mgr_get_channel_at(int ndx, core_driver_channel_config_t* chnl_cfg)
   else
   {
     chnl_cfg->chnl_type       = channel_type_analog;
-    chnl_cfg->init_val.f      = (float)cfg_mgr_get_double(node, "init_val");
-    chnl_cfg->failsafe_val.f  = (float)cfg_mgr_get_double(node, "failsafe_val");
-    chnl_cfg->min_val         = (float)cfg_mgr_get_double(node, "min_val");
-    chnl_cfg->max_val         = (float)cfg_mgr_get_double(node, "max_val");
+    chnl_cfg->init_val.f      = cfg_mgr_get_double(node, "init_val");
+    chnl_cfg->failsafe_val.f  = cfg_mgr_get_double(node, "failsafe_val");
+    chnl_cfg->min_val         = cfg_mgr_get_double(node, "min_val");
+    chnl_cfg->max_val         = cfg_mgr_get_double(node, "max_val");
   }
 
   str = cfg_mgr_get_str(node, "chnl_dir");
@@ -825,7 +825,7 @@ cfg_mgr_get_alarm_at(int ndx, core_driver_alarm_config_t* alm_cfg)
   }
   else
   {
-    alm_cfg->set_point.f = (float)cfg_mgr_get_double(node, "set_point");
+    alm_cfg->set_point.f = cfg_mgr_get_double(node, "set_point");
   }
 
   str = cfg_mgr_get_str(node, "severity");
